delete and rename files for cypress tests
https://coveord.atlassian.net/browse/KIT-903
