--- packages/atomic/cypress/integration/facets-v1/facet/facet-actions.ts
@@ -1,15 +1,0 @@
-import {TestFixture, addTag, TagProps} from '../../../fixtures/test-fixture';
-import {FacetSelectors} from './facet-selectors';
-
-export const defaultNumberOfValues = 8;
-export const label = 'The Authors';
-export const field = 'author';
-
-export const addFacet =
-  (props: TagProps = {}) =>
-  (env: TestFixture) =>
-    addTag(env, 'atomic-facet-v1', props);
-
-export function selectIdleBoxValueAt(index: number) {
-  FacetSelectors.idleBoxValue().eq(index).click();
-}
