+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-actions.ts
@@ -1,4 +1,4 @@
-import {setUpPage} from '../../../utils/setupComponent';
+import {addTag, TagProps, TestFixture} from '../../../fixtures/test-fixture';
 import {CategoryFacetSelectors} from './category-facet-selectors';
 
 export const canadaHierarchy = [
@@ -9,41 +9,32 @@ export const canadaHierarchy = [
 ];
 export const canadaHierarchyIndex = [0, 1, 0, 4];
 export const togoHierarchy = ['Africa', 'Togo', 'Lome'];
-export const quebecHierarchy = ['North America', 'Canada', 'Quebec', 'Quebec'];
 export const hierarchicalField = 'geographicalhierarchy';
 export const defaultNumberOfValues = 5;
 
 export interface CategoryFacetSetupOptions {
   field: string;
   attributes: string;
-  executeFirstSearch: boolean;
   withResultList: boolean;
 }
 
-// TODO: adapt to new setup
-export function setupCategoryFacet(
-  options: Partial<CategoryFacetSetupOptions> = {}
-) {
-  const setupOptions: CategoryFacetSetupOptions = {
-    attributes: '',
-    executeFirstSearch: true,
-    field: hierarchicalField,
-    withResultList: false,
-    ...options,
+export const addCategoryFacet =
+  (props: TagProps = {}, withResultList = false) =>
+  (env: TestFixture) => {
+    addTag(env, 'atomic-breadcrumb-manager', {});
+    addTag(env, 'atomic-category-facet-v1', {
+      field: hierarchicalField,
+      label: 'Atlas',
+      'number-of-values': defaultNumberOfValues,
+      ...props,
+    });
+    withResultList && addTag(env, 'atomic-result-list', {});
   };
-  setUpPage(
-    `<atomic-breadcrumb-manager></atomic-breadcrumb-manager>     
-      <atomic-category-facet field="${setupOptions.field}" label="Atlas" ${
-      setupOptions.attributes
-    }></atomic-category-facet>
-      ${
-        setupOptions.withResultList &&
-        '<atomic-result-list></atomic-result-list>'
-      }`,
-    setupOptions.executeFirstSearch
-  );
-}
 
 export function selectChildValueAt(index: number) {
   CategoryFacetSelectors.childValue().eq(index).click();
+}
+
+export function selectSearchResultAt(index: number) {
+  CategoryFacetSelectors.searchResult().eq(index).click();
 }
