+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-assertions.ts
@@ -3,27 +3,16 @@ import {
   doSortOccurences,
 } from '../../../utils/componentUtils';
 import {RouteAlias} from '../../../utils/setupComponent';
-import {ComponentErrorSelectors} from '../../component-error-selectors';
 import {ResultListSelectors} from '../../result-list/result-list-selectors';
 import {
   CategoryFacetSelectors,
-  categoryFacetComponent,
   BreadcrumbSelectors,
 } from './category-facet-selectors';
 import {hierarchicalField} from './category-facet-actions';
-
-function should(should: boolean) {
-  return should ? 'should' : 'should not';
-}
-
-export function assertAccessibility() {
-  it('should pass accessibility tests', () => {
-    cy.checkA11y(categoryFacetComponent);
-  });
-}
+import {should} from '../../common-assertions';
 
 export function assertNumberOfChildValues(value: number) {
-  it(`should display ${value} number of child values`, () => {
+  it(`should display ${value} number of idle link values`, () => {
     if (value > 0) {
       CategoryFacetSelectors.childValue().its('length').should('eq', value);
       return;
@@ -47,60 +36,6 @@ export function assertNumberOfParentValues(value: number) {
     CategoryFacetSelectors.parentValue()
       .its('length')
       .should('eq', value - 1);
-  });
-}
-
-export function assertLabelContains(label: string) {
-  it(`should have the label ${label}`, () => {
-    CategoryFacetSelectors.label().contains(label);
-  });
-}
-
-export function assertDisplayShowMoreButton(display: boolean) {
-  it(`${should(display)} display a "Show more" button`, () => {
-    CategoryFacetSelectors.showMoreButton().should(
-      display ? 'be.visible' : 'not.exist'
-    );
-  });
-}
-
-export function assertDisplayShowLessButton(display: boolean) {
-  it(`${should(display)} display a "Show less" button`, () => {
-    CategoryFacetSelectors.showLessButton().should(
-      display ? 'be.visible' : 'not.exist'
-    );
-  });
-}
-
-export function assertDisplayClearButton(display: boolean) {
-  it(`${should(display)} display a "All Categories" button`, () => {
-    CategoryFacetSelectors.clearButton().should(
-      display ? 'be.visible' : 'not.exist'
-    );
-  });
-}
-
-export function assertContainsComponentError(display: boolean) {
-  it(`${should(display)} display an error component`, () => {
-    CategoryFacetSelectors.shadow()
-      .find(ComponentErrorSelectors.component)
-      .should(display ? 'be.visible' : 'not.exist');
-  });
-}
-
-export function assertDisplayFacet(display: boolean) {
-  it(`${should(display)} display the facet`, () => {
-    CategoryFacetSelectors.wrapper().should(
-      display ? 'be.visible' : 'not.exist'
-    );
-  });
-}
-
-export function assertDisplayPlaceholder(display: boolean) {
-  it(`${should(display)} display the placeholder`, () => {
-    CategoryFacetSelectors.placeholder().should(
-      display ? 'be.visible' : 'not.exist'
-    );
   });
 }
 
@@ -242,5 +177,24 @@ export function assertNumberOfSearchResults(numberOfResults: number) {
     cy.get(ResultListSelectors.component)
       .find(ResultListSelectors.result)
       .should('have.length', numberOfResults);
+  });
+}
+
+export function assertDisplaySearchInput(display: boolean) {
+  it(`${should(display)} display a the facet search input`, () => {
+    CategoryFacetSelectors.searchInput().should(
+      display ? 'be.visible' : 'not.exist'
+    );
+  });
+}
+
+export function assertNumberOfFacetSearchResults(value: number) {
+  it(`should display ${value} number of facet search results`, () => {
+    if (value > 0) {
+      CategoryFacetSelectors.searchResult().its('length').should('eq', value);
+      return;
+    }
+
+    CategoryFacetSelectors.searchResult().should('not.exist');
   });
 }
