+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-selectors.ts
@@ -6,25 +6,43 @@ export const BreadcrumbSelectors = {
     BreadcrumbSelectors.shadow().find('[part="breadcrumb"]'),
 };
 
-export const categoryFacetComponent = 'atomic-category-facet';
+export const categoryFacetComponent = 'atomic-category-facet-v1';
 export const CategoryFacetSelectors = {
   shadow: () => cy.get(categoryFacetComponent).shadow(),
   wrapper: () => CategoryFacetSelectors.shadow().find('[part="facet"]'),
-  label: () => CategoryFacetSelectors.shadow().find('[part="label"]'),
+  labelButton: () =>
+    CategoryFacetSelectors.shadow().find('[part="label-button"]'),
   placeholder: () =>
     CategoryFacetSelectors.shadow().find('[part="placeholder"]'),
   clearButton: () =>
-    CategoryFacetSelectors.shadow().find('[part="clear-button"]'),
+    CategoryFacetSelectors.shadow().find('[part="all-categories-button"]'),
   valueLabel: () =>
     CategoryFacetSelectors.shadow().find('[part="value-label"]'),
+  values: () => CategoryFacetSelectors.shadow().find('[part="values"]'),
+  childValue: () => CategoryFacetSelectors.shadow().find('[part="value-link"]'),
   valueCount: () =>
     CategoryFacetSelectors.shadow().find('[part="value-count"]'),
-  childValue: () => CategoryFacetSelectors.shadow().find('[part="child"]'),
-  parentValue: () => CategoryFacetSelectors.shadow().find('[part="parent"]'),
+  parentValue: () =>
+    CategoryFacetSelectors.shadow().find('[part="parent-button"]'),
   activeParentValue: () =>
     CategoryFacetSelectors.shadow().find('[part="active-parent"]'),
   showMoreButton: () =>
     CategoryFacetSelectors.shadow().find('[part="show-more"]'),
   showLessButton: () =>
     CategoryFacetSelectors.shadow().find('[part="show-less"]'),
+  searchInput: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-input"]'),
+  searchClearButton: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-clear-button"]'),
+  searchResults: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-results"]'),
+  searchResult: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-result"]'),
+  searchResultPath: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-result-path"]'),
+  moreMatches: () =>
+    CategoryFacetSelectors.shadow().find('[part="more-matches"]'),
+  noMatches: () => CategoryFacetSelectors.shadow().find('[part="no-matches"]'),
+  valueHighlight: () =>
+    CategoryFacetSelectors.shadow().find('[part="search-highlight"]'),
 };
