+++ packages/atomic/cypress/integration/facets/facet-common-actions.ts
