+++ packages/atomic/cypress/integration/facets/numeric-facet/numeric-facet-selectors.ts
