+++ packages/atomic/cypress/integration/no-results.cypress.ts
@@ -1,37 +1,42 @@
-import {
-  buildTestUrl,
-  injectComponent,
-  setUpPage,
-} from '../utils/setupComponent';
+import {TestFixture, generateComponentHTML} from '../fixtures/test-fixture';
+import {addSearchBox} from '../fixtures/test-fixture-search-box';
 import {generateAliasForSearchBox} from './search-box-selectors';
 
 describe('No Results Test Suites', () => {
-  const tag = 'atomic-no-results';
-  const component = (attributes = '') => `<${tag} ${attributes}></${tag}>`;
-  const searchBox = '<atomic-search-box></atomic-search-box>';
+  const tag = 'atomic-no-results-v1';
   const wait = 1000;
+  let env: TestFixture;
+
+  beforeEach(() => {
+    env = new TestFixture()
+      .with(addSearchBox())
+      .withElement(generateComponentHTML(tag));
+  });
 
   it('should not be visible when there are results', () => {
-    setUpPage(component());
-    cy.get('atomic-no-results').should('not.be.visible');
+    env.init();
+    cy.get(tag).should('not.be.visible');
   });
 
   it('should be visible when there are no results', () => {
-    cy.visit(buildTestUrl('q=gahaiusdhgaiuewjfsf'));
-    injectComponent(component() + searchBox);
-    cy.wait(wait);
-    cy.get('atomic-no-results').should('be.visible');
+    env.withHash('q=gahaiusdhgaiuewjfsf').init();
+    cy.get(tag).should('be.visible');
+  });
+
+  it('text content should match when there are no results', () => {
+    env.withHash('q=dsmndkndjnj').init();
+    cy.get(tag)
+      .shadow()
+      .find('div div[part="no-results"] .quotations')
+      .should('contain.text', 'dsmndkndjnj');
   });
 
   it('cancel button should not be visible when there is no history', () => {
-    cy.visit(buildTestUrl('q=gahaiusdhgaiuewjfsf'));
-    injectComponent(component());
-    cy.wait(wait);
-    cy.get('atomic-no-results').shadow().get('button').should('not.exist');
+    env.withHash('q=gahaiusdhgaiuewjfsf').init();
+    cy.get(tag).shadow().get('button').should('not.exist');
   });
 
   function submitNoResultsSearch() {
-    setUpPage(component() + searchBox);
     generateAliasForSearchBox();
     cy.get('@searchBoxFirstDiv')
       .find('.search-input')
@@ -41,14 +46,16 @@ describe('No Results Test Suites', () => {
   }
 
   it('cancel button should be visible when there is history', () => {
+    env.init();
     submitNoResultsSearch();
-    cy.get('atomic-no-results').shadow().find('button').should('be.visible');
+    cy.get(tag).shadow().find('button').should('be.visible');
   });
 
   it('clicking on cancel should go back in history and hide the atomic-no-results component', () => {
+    env.init();
     submitNoResultsSearch();
-    cy.get('atomic-no-results').shadow().find('button').click();
+    cy.get(tag).shadow().find('button').click();
     cy.wait(wait);
-    cy.get('atomic-no-results').should('not.be.visible');
+    cy.get(tag).should('not.be.visible');
   });
 });
